Make Qt 6 qml happy, filter out non-buttons
This makes Qt 6 QML happy by filtering out all things that aren't
RadioButtons in the children when setting up the ButtonGroups.

Because of this, I had to make it somewhat complex because Qt 5's JS
list implementation doesn't have filters. This was something added in Qt
6 -- now the buttons property is initialized on component completion,
different whether or not filter is present or not.